@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 21
-    buildToolsVersion "21.1.0"
+    compileSdkVersion 28
+    buildToolsVersion "28.0.0"
 
     defaultConfig {
         applicationId "com.megasteelx.exse"
         minSdkVersion 17
-        targetSdkVersion 21
+        targetSdkVersion 28
         versionCode 20200927
         versionName "1.0"
     }
@@ -20,10 +20,10 @@ android {
 }
 
 dependencies {
-	compile 'com.github.relish-wang:ColorPicker:0.0.2-SNAPSHOT' // 0.x.x 老版本
-	//compile 'wang.relish.widget:colorpicker:1.0.0' // 1.x.x 新版本(支持AndroidX)
-	compile 'com.leon:lfilepickerlibrary:1.8.0'
-	compile 'com.android.support:support-v4:+'
+	//compile 'com.github.relish-wang:ColorPicker:0.0.2-SNAPSHOT' // 0.x.x 老版本
+	compile 'wang.relish.widget:colorpicker:1.0.0' // 1.x.x 新版本(支持AndroidX)
+	//compile 'com.leon:lfilepickerlibrary:1.8.0'
+	compile 'androidx.drawerlayout:drawerlayout:1.0.0'
     compile fileTree(dir: 'libs', include: ['*.jar'])
 }
 
